requirements.txt
flask>=2.0.0
python-telegram-bot>=20.0
schedule>=1.2.0
python-dotenv>=1.0.0
dataclasses-json>=0.6.0
pytest>=7.0.0
requests>=2.28.0
